% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network.R
\name{.interpolate_metric}
\alias{.interpolate_metric}
\title{Spline-interpolate a metric matrix to higher temporal resolution.}
\usage{
.interpolate_metric(metric_mat, bin_mids, fine_times)
}
\arguments{
\item{metric_mat}{n_vertices x n_bins matrix}

\item{bin_mids}{numeric vector of original time points (length n_bins)}

\item{fine_times}{numeric vector of target time points}
}
\value{
n_vertices x length(fine_times) matrix
}
\description{
Spline-interpolate a metric matrix to higher temporal resolution.
}
\keyword{internal}
