% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network.R
\name{extract_snapshot}
\alias{extract_snapshot}
\title{Extract Static Network Snapshot at a Time Point}
\usage{
extract_snapshot(x, at)
}
\arguments{
\item{x}{A \code{temporal_network} object.}

\item{at}{Numeric. Time point at which to extract the snapshot.}
}
\value{
An \code{igraph} object with edges active at time \code{at}.
}
\description{
Extracts a static igraph network containing all edges active at a given
time point.
}
