% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_registry.R
\name{get_estimator}
\alias{get_estimator}
\title{Retrieve a Registered Estimator}
\usage{
get_estimator(name)
}
\arguments{
\item{name}{Character. Name of the estimator to retrieve.}
}
\value{
A list with elements \code{fn}, \code{description}, \code{directed}.
}
\description{
Retrieve a registered network estimator by name.
}
\seealso{
\code{\link{register_estimator}}, \code{\link{list_estimators}}
}
