% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pcor_network.R
\name{plot.pcor_network}
\alias{plot.pcor_network}
\title{Plot Method for Partial Correlation Network}
\usage{
\method{plot}{pcor_network}(x, ...)
}
\arguments{
\item{x}{A \code{pcor_network} object.}

\item{...}{Additional arguments passed to \code{cograph::tplot()}.}
}
\description{
Plots the partial correlation network using \code{cograph::tplot()}.
Requires the \pkg{cograph} package to be installed.
}
