% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_network.R
\name{plot.saqr_network}
\alias{plot.saqr_network}
\title{Plot Method for saqr_network}
\usage{
\method{plot}{saqr_network}(x, ...)
}
\arguments{
\item{x}{A \code{saqr_network} object.}

\item{...}{Additional arguments passed to \code{cograph::splot()}.}
}
\description{
Plots the network using \code{cograph::splot()}.
Requires the \pkg{cograph} package to be installed.
}
