% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_network.R
\name{plot.saqr_network_ml}
\alias{plot.saqr_network_ml}
\title{Plot Method for saqr_network_ml}
\usage{
\method{plot}{saqr_network_ml}(x, ...)
}
\arguments{
\item{x}{A \code{saqr_network_ml} object.}

\item{...}{Additional arguments passed to \code{cograph::splot()}.}
}
\description{
Plots between and within networks side by side using
\code{cograph::splot()}.
}
