% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvar.R
\name{print.mlvar_result}
\alias{print.mlvar_result}
\title{Print method for mlvar_result}
\usage{
\method{print}{mlvar_result}(x, ...)
}
\arguments{
\item{x}{An \code{mlvar_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for mlvar_result
}
