% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{print.netobject}
\alias{print.netobject}
\title{Print Method for Network Object}
\usage{
\method{print}{netobject}(x, ...)
}
\arguments{
\item{x}{A \code{netobject}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for Network Object
}
