% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{print.netobject_ml}
\alias{print.netobject_ml}
\title{Print Method for Multilevel Network Object}
\usage{
\method{print}{netobject_ml}(x, ...)
}
\arguments{
\item{x}{A \code{netobject_ml}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for Multilevel Network Object
}
