% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{print.psych_network}
\alias{print.psych_network}
\title{Print Method for Psychological Network}
\usage{
\method{print}{psych_network}(x, ...)
}
\arguments{
\item{x}{A \code{psych_network} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for Psychological Network
}
