% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{print.saqr_permutation}
\alias{print.saqr_permutation}
\title{Print Method for saqr_permutation}
\usage{
\method{print}{saqr_permutation}(x, ...)
}
\arguments{
\item{x}{A \code{saqr_permutation} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for saqr_permutation
}
