% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_reliability.R
\name{print.tna_reliability_comparison}
\alias{print.tna_reliability_comparison}
\title{Print Method for TNA Reliability Comparison}
\usage{
\method{print}{tna_reliability_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{tna_reliability_comparison} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for TNA Reliability Comparison
}
