% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_tna.R
\name{print.tna_velocity}
\alias{print.tna_velocity}
\title{Print Method for Velocity TNA Object}
\usage{
\method{print}{tna_velocity}(x, ...)
}
\arguments{
\item{x}{A \code{tna_velocity} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for Velocity TNA Object
}
