% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_registry.R
\name{remove_estimator}
\alias{remove_estimator}
\title{Remove a Registered Estimator}
\usage{
remove_estimator(name)
}
\arguments{
\item{name}{Character. Name of the estimator to remove.}
}
\value{
Invisible \code{NULL}.
}
\description{
Remove a network estimator from the registry.
}
\seealso{
\code{\link{register_estimator}}, \code{\link{list_estimators}}
}
