% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_mean}
\alias{safe_mean}
\title{Safe Mean}
\usage{
safe_mean(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Mean value or NA if vector is empty.
}
\description{
Calculate mean with handling for empty vectors.
}
\keyword{internal}
