% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_median}
\alias{safe_median}
\title{Safe Median}
\usage{
safe_median(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Median value or NA if vector is empty.
}
\description{
Calculate median with handling for empty vectors.
}
\keyword{internal}
