% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvar.R
\name{summary.mlvar_result}
\alias{summary.mlvar_result}
\title{Summary method for mlvar_result}
\usage{
\method{summary}{mlvar_result}(object, ...)
}
\arguments{
\item{object}{An \code{mlvar_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for mlvar_result
}
