% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_network.R
\name{summary.saqr_bootstrap}
\alias{summary.saqr_bootstrap}
\title{Summary Method for saqr_bootstrap}
\usage{
\method{summary}{saqr_bootstrap}(object, ...)
}
\arguments{
\item{object}{A \code{saqr_bootstrap} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data frame with edge-level bootstrap statistics.
}
\description{
Summary Method for saqr_bootstrap
}
