% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tna_networks.R
\name{generate_sequence_data}
\alias{generate_sequence_data}
\title{Generate Sequence Data (Deprecated)}
\usage{
generate_sequence_data(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been renamed to \code{\link{generate_tna_datasets}}.
Please use \code{generate_tna_datasets()} instead.
}
\keyword{internal}
