% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_names.R
\docType{data}
\name{GLOBAL_NAMES}
\alias{GLOBAL_NAMES}
\title{Global Names Dataset}
\format{
A character vector of 100 names from various cultures and regions.
}
\usage{
GLOBAL_NAMES
}
\description{
A collection of 100 short, diverse names from around the world for use in
social network simulations.
}
\details{
Names are selected to be short (typically 3-7 characters) and represent
diverse cultural backgrounds including:
\itemize{
\item European (Western, Eastern, Nordic)
\item Asian (East Asian, South Asian, Southeast Asian, Middle Eastern)
\item African
\item Latin American
\item Oceanian
}
}
\examples{
# View all names
GLOBAL_NAMES

# Sample 10 random names
sample(GLOBAL_NAMES, 10)

# Use in network simulation
simulate_edge_list(n_nodes = 20)

}
\keyword{datasets}
